SDL Documentation (Draft)
My current documentation for any of the MacOS building you may want to do.
May expand to universal platform loading within the .json file of mine. Another thing I may want to consider is a CMake variation, just so that I cover most of the build bases, and so that you can jump from OS to OS without much downtime